lib: split between general and task-specific allocation code
